Split files into separate classes
